attempt to restore withExpiringActivity
